+++ CHANGELOG.md
@@ -1,3 +1,4 @@
-## [0.0.1] - TODO: Add release date.
+## [0.0.1] - 2020-11-06
 
-* TODO: Describe initial release.
+* This release contains initial version of this release where user can create a custom list row widget
+with three column layout.